test: co-locate test files alongside their source modules
