+++ src/layout/sidebar/settings-sidebar.test.tsx
@@ -3,15 +3,17 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import type { AuthClient } from '@/contexts'
+import { SignInModalProvider } from '@/contexts'
+import { SidebarProvider } from '@/components/ui/sidebar'
 import { resetTestDatabase, setupTestDatabase, teardownTestDatabase } from '@/dal/test-utils'
 import { createMockAuthClient } from '@/test-utils/auth-client'
 import { createTestProvider } from '@/test-utils/test-provider'
 import '@testing-library/jest-dom'
 import { cleanup, render, screen } from '@testing-library/react'
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it } from 'bun:test'
+import { MemoryRouter } from 'react-router'
 import type { ReactNode } from 'react'
-import { MemoryRouter, Route, Routes } from 'react-router'
-import AgentsSettingsPage from '../index'
+import { SettingsSidebarContent } from './settings-sidebar'
 
 const anonSession = {
   user: { id: 'anon-1', email: '', name: '', isAnonymous: true },
@@ -21,27 +23,27 @@ const authedSession = {
   user: { id: 'user-1', email: 'a@b.com', name: 'Alice', isAnonymous: false },
 }
 
-const settingsIndexMarker = 'settings-index-marker'
-
-const renderPage = (authClient: AuthClient, isStandalone: () => boolean) => {
+const renderSidebar = (authClient: AuthClient, isStandalone: () => boolean) => {
   const TestProvider = createTestProvider({ authClient })
   const Wrapper = ({ children }: { children: ReactNode }) => (
     <TestProvider>
-      <MemoryRouter initialEntries={['/settings/agents']}>
-        <Routes>
-          <Route path="/settings/agents" element={children} />
-          <Route path="/settings" element={<div data-testid={settingsIndexMarker} />} />
-        </Routes>
-      </MemoryRouter>
+      <SignInModalProvider>
+        <MemoryRouter initialEntries={['/settings']}>
+          <SidebarProvider>{children}</SidebarProvider>
+        </MemoryRouter>
+      </SignInModalProvider>
     </TestProvider>
   )
-  return render(<AgentsSettingsPage isStandalone={isStandalone} />, { wrapper: Wrapper })
+  return render(
+    <SettingsSidebarContent onBackClick={() => {}} onSettingsNavigate={() => {}} isStandalone={isStandalone} />,
+    { wrapper: Wrapper },
+  )
 }
 
 const onTauri = () => true
 const offTauri = () => false
 
-describe('AgentsSettingsPage — hidden state guard', () => {
+describe('SettingsSidebarContent — Agents entry visibility', () => {
   beforeAll(async () => {
     await setupTestDatabase()
   })
@@ -60,43 +62,40 @@ describe('AgentsSettingsPage — hidden state guard', () => {
     localStorage.clear()
   })
 
-  it('redirects to /settings for anonymous users when the proxy is effectively on (web)', () => {
+  it('hides the Agents entry for anonymous users when the proxy is effectively on (web)', () => {
     const authClient = createMockAuthClient({ session: anonSession })
-    renderPage(authClient, offTauri)
+    renderSidebar(authClient, offTauri)
 
-    expect(screen.getByTestId(settingsIndexMarker)).toBeInTheDocument()
     expect(screen.queryByText('Agents')).not.toBeInTheDocument()
   })
 
-  it('redirects for anonymous users on Tauri Connected (proxy_enabled=true)', () => {
+  it('hides the Agents entry for anonymous users on Tauri Connected (proxy_enabled=true)', () => {
     localStorage.setItem('proxy_enabled', 'true')
     const authClient = createMockAuthClient({ session: anonSession })
-    renderPage(authClient, onTauri)
+    renderSidebar(authClient, onTauri)
 
-    expect(screen.getByTestId(settingsIndexMarker)).toBeInTheDocument()
+    expect(screen.queryByText('Agents')).not.toBeInTheDocument()
   })
 
-  it('renders the page for anonymous users on Tauri Standalone (proxy off)', () => {
+  it('shows the Agents entry for anonymous users on Tauri Standalone (proxy off)', () => {
+    // localStorage has no `proxy_enabled` — defaults to false on Tauri.
     const authClient = createMockAuthClient({ session: anonSession })
-    renderPage(authClient, onTauri)
+    renderSidebar(authClient, onTauri)
 
-    expect(screen.queryByTestId(settingsIndexMarker)).not.toBeInTheDocument()
-    expect(screen.getByRole('button', { name: /add custom agent/i })).toBeInTheDocument()
+    expect(screen.getByText('Agents')).toBeInTheDocument()
   })
 
-  it('renders the page for authenticated users behind the proxy (web)', () => {
+  it('shows the Agents entry for authenticated users behind the proxy (web)', () => {
     const authClient = createMockAuthClient({ session: authedSession })
-    renderPage(authClient, offTauri)
+    renderSidebar(authClient, offTauri)
 
-    expect(screen.queryByTestId(settingsIndexMarker)).not.toBeInTheDocument()
-    expect(screen.getByRole('button', { name: /add custom agent/i })).toBeInTheDocument()
+    expect(screen.getByText('Agents')).toBeInTheDocument()
   })
 
-  it('renders the page for authenticated users on Tauri Standalone', () => {
+  it('shows the Agents entry for authenticated users on Tauri Standalone (proxy off)', () => {
     const authClient = createMockAuthClient({ session: authedSession })
-    renderPage(authClient, onTauri)
+    renderSidebar(authClient, onTauri)
 
-    expect(screen.queryByTestId(settingsIndexMarker)).not.toBeInTheDocument()
-    expect(screen.getByRole('button', { name: /add custom agent/i })).toBeInTheDocument()
+    expect(screen.getByText('Agents')).toBeInTheDocument()
   })
 })
